separate data manager to local and network classes #3
