+++ HomeForSale/Network/LocalDataManager.swift
@@ -6,11 +6,12 @@
 //
 
 import Foundation
+import UIKit
 
 final class LocalDataManager: DataManagerProtocol {
-    func fetchJson(_ path: String, completion: @escaping (Result<Data, Error>) -> Void) {
+    func fetchJson(forType type: RequestType, completion: @escaping (Result<Data, Error>) -> Void) {
         do {
-            if let bundlePath = Bundle.main.path(forResource: path,
+            if let bundlePath = Bundle.main.path(forResource: type.filePath,
                                                  ofType: "json"),
                let jsonData = try String(contentsOfFile: bundlePath).data(using: .utf8) {
                 completion(.success(jsonData))
@@ -22,6 +23,14 @@ final class LocalDataManager: DataManagerProtocol {
     }
 
     func fetchDataImage(_ path: String, completion: @escaping (Result<Data, Error>) -> Void) {
+        let image = UIImage(named: "houseImg")
+        let imageData = image?.pngData()
+
+        if let data = imageData {
+            completion(.success(data))
+        } else {
+            completion(.failure(NetworkError.badURL))
+        }
 
     }
